refactor: DTO 패키지 분리 및 불필요한 필드 제거 (#3)
- 농어촌 api 관련 DTO rural 패키지로 분리
- 테마관광 api 관련 DTO theme 패키지로 분리
- 농가맛집 api 관련 DTO farm 패키지로 분리
- 농가맛집 리스트 관련 DTO farmList~ 형식으로 rename
- FarmListItem 불필요한 필드 제거
- 공통으로 사용하는 DTO는 common 패키지로 분리
