+++ src/main/java/com/chocobi/leafy/place/dto/common/Response.java
@@ -1,4 +1,4 @@
-package com.chocobi.leafy.place.dto;
+package com.chocobi.leafy.place.dto.common;
 
 import lombok.Data;
