+++ src/main/java/com/chocobi/leafy/place/dto/farm/FarmListApiResponse.java
@@ -1,5 +1,6 @@
-package com.chocobi.leafy.place.dto;
+package com.chocobi.leafy.place.dto.farm;
 
+import com.chocobi.leafy.place.dto.common.Header;
 import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
 import jakarta.xml.bind.annotation.XmlRootElement;
@@ -8,7 +9,7 @@ import lombok.Data;
 @Data
 @XmlRootElement(name = "response")
 @XmlAccessorType(XmlAccessType.FIELD)
-public class FarmApiResponse {
+public class FarmListApiResponse {
     private Header header;
-    private FarmXmlBody body;
+    private FarmListBody body;
 }
