+++ src/main/java/com/chocobi/leafy/place/dto/farm/FarmListItem.java
@@ -1,0 +1,10 @@
+package com.chocobi.leafy.place.dto.farm;
+
+import lombok.Data;
+
+@Data
+public class FarmListItem {
+    private String cntntsNo;    //콘텐츠 번호
+    private String cntntsSj;    //제목
+    private String thumbImgUrl; //썸네일
+}
